CRUD BASE LISTOS, solo falta empleado

+++ 911_RD/911_RD/CIUDADES.cs
@@ -20,7 +20,7 @@ namespace _911_RD
             this.DIRECCIONES = new HashSet<DIRECCIONES>();
         }
     
-        public int id_ciudad { get; set; }
+        public double id_ciudad { get; set; }
         public int id_provincia { get; set; }
         public string ciudad { get; set; }
